only decrement counter when there aren't planks in the inventory
this is shockingly difficult since it's not readily available what
planks will be used. As a result, the code now waits for the
construction animation to finish, and then decrements. It's pretty jank,
but without some extra parsing for the current Mahogany Homes tier it's
the only way I have.