cap xp readout to 8000
Turns out the game will track up to 32k xp for each relic, as the varbit
is 16 bits instead of the minimum of 14 bits. This was probably done to
make aligning them into 32 bit ints easier, but it's odd that they
didn't cap it server side. Maybe it was to save processing time on the
server?