make sure that the overlay gets built after logging out and in
it seems that widgets don't get cleared until quite late into loading.
this causes the overlay building to think it worked with no problems,
when in actuality it hasn't. By forcing the GameTick handler to build
the overlay, we can make sure that the overlay is actually, truly, and
properly built, as gameticks don't seem to start happening until after
the widgets have been cleaned and fixed.