Stop cursors from overlapping with tooltips
I don't think addFront was a thing when I first made this, but that
along with being able to use an image as a tooltip means it's relatively
simple to make a spacer tooltip to push the others down an arbitrary
amount.
The main issue was ordering so this overlay is rendered after
the Mouse Tooltips one, as otherwise it would addFront its own tooltip
after mine. The downside to changing the layer is that tooltips
themselves are rendered above the cursor, which makes the spacer
essential.