clear tracked counts when switching profiles
I think this was the cause of the weird "you have 0 left" issues that
happened a while back, as if you switched from an account that had a
pretty full book onto an account with an empty/less filled book, the
plugin would see the negative difference and assume you teleported.

I think.