only request default states if the player in question is in the party
This was likely the other half of the problem the prior commit fixed.
In a busy area, a lot of players will have been checked, and most won't
have a default state since they're not in your party. Each one of those
will have caused a UserSync with the old code, which is just really bad.