Fix cursor rendering behind minimenu
So this was a fun one. To fix the overlap issue, I changed the layer of
the overlay. This was to make it render before the Tooltip overlay, but
after Mouse Tooltips, such as to insert a tooltip at the front. This
worked, but this new layer meant that the minimenu was rendered on top.
Turns out that it's actually impossible to have a single overlay add to
the actual front of the tooltips and render above the minimenu.
So now I have 2. Will this break in a different way? Probably. I've no
idea how it would break, but at this point it feels nearly guaranteed.