remove deprecated API usage
wow this was way more of a pain to do than it really should have been.
Cryptic error on the hub build log that didn't say what was actually
wrong, references to files that don't actually really exist, but are
built at compile(?) time, a not 1-to-1 translation between old and new
code with no indication on the deprecated code about how to do it.

yikes